gộp fragment likedRoom thành 1, thêm nút rented cho details
