+++ app/src/main/res/layout/activity_details.xml
@@ -49,7 +49,23 @@
             android:button="@drawable/heart_checkbox_selector"
             app:buttonTint="@color/white"
             android:buttonTintMode="multiply"
-            android:layout_marginRight="10dp"/>
+            android:layout_marginRight="10dp"
+            android:visibility="gone"/>
+        <com.google.android.material.checkbox.MaterialCheckBox
+            android:id="@+id/cbIsRented"
+            android:layout_width="wrap_content"
+            android:layout_height="match_parent"
+            android:text="Đã cho thuê"
+            android:layoutDirection="rtl"
+            android:layout_gravity="right"
+            android:layout_marginRight="10dp"
+            android:textSize="17dp"
+            android:textColor="@color/white"
+            app:buttonIconTint="@color/white"
+            android:checked="true"
+            >
+
+        </com.google.android.material.checkbox.MaterialCheckBox>
     </com.google.android.material.appbar.MaterialToolbar>
 
     <ScrollView
@@ -760,8 +776,7 @@
             android:layout_height="wrap_content"
             android:layout_marginHorizontal="20dp"
             android:layout_weight="1"
-            android:visibility="gone">
-
+            >
             <com.google.android.material.button.MaterialButton
                 android:id="@+id/btnToDelete"
                 android:layout_width="0dp"
@@ -798,6 +813,8 @@
                 app:iconSize="22dp"
                 app:cornerRadius="10dp"
                 />
+
+
         </LinearLayout>
