--- app/src/main/res/layout/fragment_liked_empty.xml
@@ -1,29 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".FragmentFilter.LikedEmptyFragment"
-    android:background="@drawable/layout_bg">
-
-    <!-- TODO: Update blank fragment layout -->
-    <ImageView
-        android:id="@+id/face_nothing"
-        android:layout_width="150dp"
-        android:layout_height="150dp"
-        android:src="@mipmap/ic_face_foreground"
-        android:layout_gravity="center"
-
-        />
-    <TextView
-        android:id="@+id/text"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="Bạn chưa có phòng nào thích"
-        android:layout_gravity="center"
-        android:layout_marginTop="100dp"
-        android:textSize="15sp"
-        />
-
-
-</FrameLayout>
