--- app/src/main/res/layout/fragment_liked_rooms.xml
@@ -1,25 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".FragmentFilter.FragmentLikedRooms"
-    android:background="@drawable/layout_bg">
-
-    <!-- TODO: Update blank fragment layout -->
-
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/rcv_LikedRooms"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:layout_marginBottom="10dp"
-        android:layout_marginTop="10dp"
-        android:layout_gravity="center"
-        android:layout_marginRight="5dp"
-
-        >
-
-    </androidx.recyclerview.widget.RecyclerView>
-
-
-</FrameLayout>
